requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# For UI
streamlit

#Gemini
langchain-google-genai

#Document Loaders
PyPDFLoader
